Extract HighlightChars into standalone component
