+++ packages/studio-base/src/components/HighlightChars.tsx
@@ -1,0 +1,30 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+// Renders the given text with the characters at the given indices wrapped in a
+// <mark> component for Fzf results. The indices are the positions of the
+// matched characters in the original string.
+//
+// Optionally, an offset can be provided to account for the fact that the search
+// string may be a substring of the original string.
+
+type Props = {
+  str: string;
+  indices: Set<number>;
+  offset?: number;
+};
+
+export function HighlightChars(props: Props): JSX.Element {
+  const { str, indices, offset = 0 } = props;
+  const chars = str.split("");
+
+  const nodes = chars.map((char, i) => {
+    if (indices.has(i + offset)) {
+      return <mark key={i}>{char}</mark>;
+    }
+    return char;
+  });
+
+  return <>{nodes}</>;
+}
