+++ packages/studio-base/src/components/TopicList.tsx
@@ -12,7 +12,6 @@ import {
   Skeleton,
   TextField,
   Typography,
-  TypographyProps,
 } from "@mui/material";
 import { Fzf, FzfResultItem } from "fzf";
 import { useMemo, useState } from "react";
@@ -20,6 +19,7 @@ import { makeStyles } from "tss-react/mui";
 
 import { DirectTopicStatsUpdater } from "@foxglove/studio-base/components/DirectTopicStatsUpdater";
 import EmptyState from "@foxglove/studio-base/components/EmptyState";
+import { HighlightChars } from "@foxglove/studio-base/components/HighlightChars";
 import {
   MessagePipelineContext,
   useMessagePipeline,
@@ -40,33 +40,6 @@ const topicToFzfResult = (item: TopicWithStats) =>
     end: 0,
   } as FzfResultItem<TopicWithStats>);
 
-const HighlightChars = ({
-  str,
-  indices,
-  color,
-  offset = 0,
-}: {
-  str: string;
-  indices: Set<number>;
-  color?: TypographyProps["color"];
-  offset?: number;
-}) => {
-  const chars = str.split("");
-
-  const nodes = chars.map((char, i) => {
-    if (indices.has(i + offset)) {
-      return (
-        <Typography component="b" key={i} variant="inherit" color={color ?? "info.main"}>
-          {char}
-        </Typography>
-      );
-    }
-    return char;
-  });
-
-  return <>{nodes}</>;
-};
-
 const useStyles = makeStyles()((theme) => ({
   appBar: {
     top: 0,
@@ -78,6 +51,10 @@ const useStyles = makeStyles()((theme) => ({
   listItem: {
     paddingRight: theme.spacing(1),
 
+    mark: {
+      color: theme.palette.info.main,
+      fontWeight: "bold",
+    },
     "&.MuiListItem-dense": {
       ".MuiListItemText-root": {
         marginTop: theme.spacing(0.5),
